Final Commit
Added all files from project 04-1 for debugging in the verifyCrops function. Errors were found with the amount, and location of the command block characters and there was a syntax error with spelling preventing it to run. At first ost had it work without noticing a syntax error because it had auto-corrected for them correctly when inputting the broken code from project 04-1.